Function rcudnn_sys::cudaDeviceCanAccessPeer

pub unsafe extern "C" fn cudaDeviceCanAccessPeer(
    canAccessPeer: *mut c_int, 
    device: c_int, 
    peerDevice: c_int
) -> cudaError_t

\brief Queries if a device may directly access a peer device's memory.

Returns in \p *canAccessPeer a value of 1 if device \p device is capable of directly accessing memory from \p peerDevice and 0 otherwise. If direct access of \p peerDevice from \p device is possible, then access may be enabled by calling ::cudaDeviceEnablePeerAccess().

\param canAccessPeer - Returned access capability \param device - Device from which allocations on \p peerDevice are to be directly accessed. \param peerDevice - Device on which the allocations to be directly accessed by \p device reside.

\return ::cudaSuccess, ::cudaErrorInvalidDevice \notefnerr \note_init_rt \note_callback

\sa ::cudaDeviceEnablePeerAccess, ::cudaDeviceDisablePeerAccess, ::cuDeviceCanAccessPeer