Function rcudnn_sys::cudaDeviceGetPCIBusId

pub unsafe extern "C" fn cudaDeviceGetPCIBusId(
    pciBusId: *mut c_char, 
    len: c_int, 
    device: c_int
) -> cudaError_t

\brief Returns a PCI Bus Id string for the device

Returns an ASCII string identifying the device \p dev in the NULL-terminated string pointed to by \p pciBusId. \p len specifies the maximum length of the string that may be returned.

\param pciBusId - Returned identifier string for the device in the following format [domain]:[bus]:[device].[function] where \p domain, \p bus, \p device, and \p function are all hexadecimal values. pciBusId should be large enough to store 13 characters including the NULL-terminator.

\param len - Maximum length of string to store in \p name

\param device - Device to get identifier string for

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorInvalidDevice \notefnerr \note_init_rt \note_callback

\sa ::cudaDeviceGetByPCIBusId, ::cuDeviceGetPCIBusId