Function rcudnn_sys::cudaEventQuery

pub unsafe extern "C" fn cudaEventQuery(event: cudaEvent_t) -> cudaError_t

\brief Queries an event's status

Queries the status of all work currently captured by \p event. See ::cudaEventRecord() for details on what is captured by an event.

Returns ::cudaSuccess if all captured work has been completed, or ::cudaErrorNotReady if any captured work is incomplete.

For the purposes of Unified Memory, a return value of ::cudaSuccess is equivalent to having called ::cudaEventSynchronize().

\param event - Event to query

\return ::cudaSuccess, ::cudaErrorNotReady, ::cudaErrorInvalidValue, ::cudaErrorInvalidResourceHandle, ::cudaErrorLaunchFailure \notefnerr \note_init_rt \note_callback

\sa \ref ::cudaEventCreate(cudaEvent_t*) "cudaEventCreate (C API)", ::cudaEventCreateWithFlags, ::cudaEventRecord, ::cudaEventSynchronize, ::cudaEventDestroy, ::cudaEventElapsedTime, ::cuEventQuery