Function rcudnn_sys::cudaEventSynchronize

pub unsafe extern "C" fn cudaEventSynchronize(event: cudaEvent_t) -> cudaError_t

\brief Waits for an event to complete

Waits until the completion of all work currently captured in \p event. See ::cudaEventRecord() for details on what is captured by an event.

Waiting for an event that was created with the ::cudaEventBlockingSync flag will cause the calling CPU thread to block until the event has been completed by the device. If the ::cudaEventBlockingSync flag has not been set, then the CPU thread will busy-wait until the event has been completed by the device.

\param event - Event to wait for

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorInvalidResourceHandle, ::cudaErrorLaunchFailure \notefnerr \note_init_rt \note_callback

\sa \ref ::cudaEventCreate(cudaEvent_t*) "cudaEventCreate (C API)", ::cudaEventCreateWithFlags, ::cudaEventRecord, ::cudaEventQuery, ::cudaEventDestroy, ::cudaEventElapsedTime, ::cuEventSynchronize