Function rcudnn_sys::cudaGraphAddMemcpyNode

pub unsafe extern "C" fn cudaGraphAddMemcpyNode(
    pGraphNode: *mut cudaGraphNode_t, 
    graph: cudaGraph_t, 
    pDependencies: *const cudaGraphNode_t, 
    numDependencies: usize, 
    pCopyParams: *const cudaMemcpy3DParms
) -> cudaError_t

\brief Creates a memcpy node and adds it to a graph

Creates a new memcpy node and adds it to \p graph with \p numDependencies dependencies specified via \p pDependencies. It is possible for \p numDependencies to be 0, in which case the node will be placed at the root of the graph. \p pDependencies may not have any duplicate entries. A handle to the new node will be returned in \p pGraphNode.

When the graph is launched, the node will perform the memcpy described by \p pCopyParams. See ::cudaMemcpy3D() for a description of the structure and its restrictions.

Memcpy nodes have some additional restrictions with regards to managed memory, if the system contains at least one device which has a zero value for the device attribute ::cudaDevAttrConcurrentManagedAccess.

\param pGraphNode - Returns newly created node \param graph - Graph to which to add the node \param pDependencies - Dependencies of the node \param numDependencies - Number of dependencies \param pCopyParams - Parameters for the memory copy

\return ::cudaSuccess, ::cudaErrorInvalidValue \note_graph_thread_safety \notefnerr \note_init_rt \note_callback

\sa ::cudaMemcpy3D, ::cudaGraphMemcpyNodeGetParams, ::cudaGraphMemcpyNodeSetParams, ::cudaGraphCreate, ::cudaGraphDestroyNode, ::cudaGraphAddChildGraphNode, ::cudaGraphAddEmptyNode, ::cudaGraphAddKernelNode, ::cudaGraphAddHostNode, ::cudaGraphAddMemsetNode