Function rcudnn_sys::cudaGraphExecKernelNodeSetParams

pub unsafe extern "C" fn cudaGraphExecKernelNodeSetParams(
    hGraphExec: cudaGraphExec_t, 
    node: cudaGraphNode_t, 
    pNodeParams: *const cudaKernelNodeParams
) -> cudaError_t

\brief Sets the parameters for a kernel node in the given graphExec

Sets the parameters of a kernel node in an executable graph \p hGraphExec. The node is identified by the corresponding node \p node in the non-executable graph, from which the executable graph was instantiated.

\p node must not have been removed from the original graph. The \p func field of \p nodeParams cannot be modified and must match the original value. All other values can be modified.

The modifications only affect future launches of \p hGraphExec. Already enqueued or running launches of \p hGraphExec are not affected by this call. \p node is also not modified by this call.

\param hGraphExec - The executable graph in which to set the specified node \param node - kernel node from the graph from which graphExec was instantiated \param pNodeParams - Updated Parameters to set

\return ::cudaSuccess, ::cudaErrorInvalidValue, \note_graph_thread_safety \notefnerr \note_init_rt \note_callback

\sa ::cudaGraphAddKernelNode, ::cudaGraphKernelNodeSetParams, ::cudaGraphExecMemcpyNodeSetParams, ::cudaGraphExecMemsetNodeSetParams, ::cudaGraphExecHostNodeSetParams, ::cudaGraphExecChildGraphNodeSetParams, ::cudaGraphExecEventRecordNodeSetEvent, ::cudaGraphExecEventWaitNodeSetEvent, ::cudaGraphExecExternalSemaphoresSignalNodeSetParams, ::cudaGraphExecExternalSemaphoresWaitNodeSetParams, ::cudaGraphExecUpdate, ::cudaGraphInstantiate