Function rcudnn_sys::cudaGraphExecMemsetNodeSetParams

pub unsafe extern "C" fn cudaGraphExecMemsetNodeSetParams(
    hGraphExec: cudaGraphExec_t, 
    node: cudaGraphNode_t, 
    pNodeParams: *const cudaMemsetParams
) -> cudaError_t

\brief Sets the parameters for a memset node in the given graphExec.

Updates the work represented by \p node in \p hGraphExec as though \p node had contained \p pNodeParams at instantiation. \p node must remain in the graph which was used to instantiate \p hGraphExec. Changed edges to and from \p node are ignored.

The destination memory in \p pNodeParams must be allocated from the same context as the original destination memory. Both the instantiation-time memory operand and the memory operand in \p pNodeParams must be 1-dimensional. Zero-length operations are not supported.

The modifications only affect future launches of \p hGraphExec. Already enqueued or running launches of \p hGraphExec are not affected by this call. \p node is also not modified by this call.

Returns cudaErrorInvalidValue if the memory operand’s mappings changed or either the original or new memory operand are multidimensional.

\param hGraphExec - The executable graph in which to set the specified node \param node - Memset node from the graph which was used to instantiate graphExec \param pNodeParams - Updated Parameters to set

\return ::cudaSuccess, ::cudaErrorInvalidValue, \note_graph_thread_safety \notefnerr \note_init_rt \note_callback

\sa ::cudaGraphAddMemsetNode, ::cudaGraphMemsetNodeSetParams ::cudaGraphInstantiate ::cudaGraphExecKernelNodeSetParams ::cudaGraphExecMemcpyNodeSetParams ::cudaGraphExecHostNodeSetParams