Function rcudnn_sys::cudaGraphGetEdges

pub unsafe extern "C" fn cudaGraphGetEdges(
    graph: cudaGraph_t, 
    from: *mut cudaGraphNode_t, 
    to: *mut cudaGraphNode_t, 
    numEdges: *mut usize
) -> cudaError_t

\brief Returns a graph’s dependency edges

Returns a list of \p graph’s dependency edges. Edges are returned via corresponding indices in \p from and \p to; that is, the node in \p to[i] has a dependency on the node in \p from[i]. \p from and \p to may both be NULL, in which case this function only returns the number of edges in \p numEdges. Otherwise, \p numEdges entries will be filled in. If \p numEdges is higher than the actual number of edges, the remaining entries in \p from and \p to will be set to NULL, and the number of edges actually returned will be written to \p numEdges.

\param graph - Graph to get the edges from \param from - Location to return edge endpoints \param to - Location to return edge endpoints \param numEdges - See description

\return ::cudaSuccess, ::cudaErrorInvalidValue \note_graph_thread_safety \notefnerr \note_init_rt \note_callback

\sa ::cudaGraphGetNodes, ::cudaGraphGetRootNodes, ::cudaGraphAddDependencies, ::cudaGraphRemoveDependencies, ::cudaGraphNodeGetDependencies, ::cudaGraphNodeGetDependentNodes