Function rcudnn_sys::cudaGraphGetRootNodes

pub unsafe extern "C" fn cudaGraphGetRootNodes(
    graph: cudaGraph_t, 
    pRootNodes: *mut cudaGraphNode_t, 
    pNumRootNodes: *mut usize
) -> cudaError_t

\brief Returns a graph’s root nodes

Returns a list of \p graph’s root nodes. \p pRootNodes may be NULL, in which case this function will return the number of root nodes in \p pNumRootNodes. Otherwise, \p pNumRootNodes entries will be filled in. If \p pNumRootNodes is higher than the actual number of root nodes, the remaining entries in \p pRootNodes will be set to NULL, and the number of nodes actually obtained will be returned in \p pNumRootNodes.

\param graph - Graph to query \param pRootNodes - Pointer to return the root nodes \param pNumRootNodes - See description

\return ::cudaSuccess, ::cudaErrorInvalidValue \note_graph_thread_safety \notefnerr \note_init_rt \note_callback

\sa ::cudaGraphCreate, ::cudaGraphGetNodes, ::cudaGraphGetEdges, ::cudaGraphNodeGetType, ::cudaGraphNodeGetDependencies, ::cudaGraphNodeGetDependentNodes