Function rcudnn_sys::cudaGraphicsResourceGetMappedMipmappedArray

pub unsafe extern "C" fn cudaGraphicsResourceGetMappedMipmappedArray(
    mipmappedArray: *mut cudaMipmappedArray_t, 
    resource: cudaGraphicsResource_t
) -> cudaError_t

\brief Get a mipmapped array through which to access a mapped graphics resource.

Returns in \p *mipmappedArray a mipmapped array through which the mapped graphics resource \p resource may be accessed. The value set in \p mipmappedArray may change every time that \p resource is mapped.

If \p resource is not a texture then it cannot be accessed via an array and ::cudaErrorUnknown is returned. If \p resource is not mapped then ::cudaErrorUnknown is returned.

\param mipmappedArray - Returned mipmapped array through which \p resource may be accessed \param resource - Mapped resource to access

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorInvalidResourceHandle, ::cudaErrorUnknown \notefnerr \note_init_rt \note_callback

\sa ::cudaGraphicsResourceGetMappedPointer, ::cuGraphicsResourceGetMappedMipmappedArray