Function rcudnn_sys::cudaMemRangeGetAttributes

pub unsafe extern "C" fn cudaMemRangeGetAttributes(
    data: *mut *mut c_void, 
    dataSizes: *mut usize, 
    attributes: *mut cudaMemRangeAttribute, 
    numAttributes: usize, 
    devPtr: *const c_void, 
    count: usize
) -> cudaError_t

\brief Query attributes of a given memory range.

Query attributes of the memory range starting at \p devPtr with a size of \p count bytes. The memory range must refer to managed memory allocated via ::cudaMallocManaged or declared via managed variables. The \p attributes array will be interpreted to have \p numAttributes entries. The \p dataSizes array will also be interpreted to have \p numAttributes entries. The results of the query will be stored in \p data.

The list of supported attributes are given below. Please refer to ::cudaMemRangeGetAttribute for attribute descriptions and restrictions.

::cudaMemRangeAttributeReadMostly

::cudaMemRangeAttributePreferredLocation

::cudaMemRangeAttributeAccessedBy

::cudaMemRangeAttributeLastPrefetchLocation

\param data - A two-dimensional array containing pointers to memory locations where the result of each attribute query will be written to. \param dataSizes - Array containing the sizes of each result \param attributes - An array of attributes to query (numAttributes and the number of attributes in this array should match) \param numAttributes - Number of attributes to query \param devPtr - Start of the range to query \param count - Size of the range to query

\return ::cudaSuccess, ::cudaErrorInvalidValue \notefnerr \note_init_rt \note_callback

\sa ::cudaMemRangeGetAttribute, ::cudaMemAdvise ::cudaMemPrefetchAsync, ::cuMemRangeGetAttributes