Function rcudnn_sys::cudaMemcpyArrayToArray

pub unsafe extern "C" fn cudaMemcpyArrayToArray(
    dst: cudaArray_t, 
    wOffsetDst: usize, 
    hOffsetDst: usize, 
    src: cudaArray_const_t, 
    wOffsetSrc: usize, 
    hOffsetSrc: usize, 
    count: usize, 
    kind: cudaMemcpyKind
) -> cudaError_t

\brief Copies data between host and device

\deprecated

Copies \p count bytes from the CUDA array \p src starting at \p hOffsetSrc rows and \p wOffsetSrc bytes from the upper left corner to the CUDA array \p dst starting at \p hOffsetDst rows and \p wOffsetDst bytes from the upper left corner, where \p kind specifies the direction of the copy, and must be one of ::cudaMemcpyHostToHost, ::cudaMemcpyHostToDevice, ::cudaMemcpyDeviceToHost, ::cudaMemcpyDeviceToDevice, or ::cudaMemcpyDefault. Passing ::cudaMemcpyDefault is recommended, in which case the type of transfer is inferred from the pointer values. However, ::cudaMemcpyDefault is only allowed on systems that support unified virtual addressing.

\param dst - Destination memory address \param wOffsetDst - Destination starting X offset (columns in bytes) \param hOffsetDst - Destination starting Y offset (rows) \param src - Source memory address \param wOffsetSrc - Source starting X offset (columns in bytes) \param hOffsetSrc - Source starting Y offset (rows) \param count - Size in bytes to copy \param kind - Type of transfer

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorInvalidMemcpyDirection \notefnerr \note_init_rt \note_callback

\sa ::cudaMemcpy, ::cudaMemcpy2D, ::cudaMemcpyToArray, ::cudaMemcpy2DToArray, ::cudaMemcpyFromArray, ::cudaMemcpy2DFromArray, ::cudaMemcpy2DArrayToArray, ::cudaMemcpyToSymbol, ::cudaMemcpyFromSymbol, ::cudaMemcpyAsync, ::cudaMemcpy2DAsync, ::cudaMemcpyToArrayAsync, ::cudaMemcpy2DToArrayAsync, ::cudaMemcpyFromArrayAsync, ::cudaMemcpy2DFromArrayAsync, ::cudaMemcpyToSymbolAsync, ::cudaMemcpyFromSymbolAsync, ::cuMemcpyAtoA