Function rcudnn_sys::cudaMemcpyPeer

pub unsafe extern "C" fn cudaMemcpyPeer(
    dst: *mut c_void, 
    dstDevice: c_int, 
    src: *const c_void, 
    srcDevice: c_int, 
    count: usize
) -> cudaError_t

\brief Copies memory between two devices

Copies memory from one device to memory on another device. \p dst is the base device pointer of the destination memory and \p dstDevice is the destination device. \p src is the base device pointer of the source memory and \p srcDevice is the source device. \p count specifies the number of bytes to copy.

Note that this function is asynchronous with respect to the host, but serialized with respect all pending and future asynchronous work in to the current device, \p srcDevice, and \p dstDevice (use ::cudaMemcpyPeerAsync to avoid this synchronization).

\param dst - Destination device pointer \param dstDevice - Destination device \param src - Source device pointer \param srcDevice - Source device \param count - Size of memory copy in bytes

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorInvalidDevice \notefnerr \note_sync \note_init_rt \note_callback

\sa ::cudaMemcpy, ::cudaMemcpyAsync, ::cudaMemcpyPeerAsync, ::cudaMemcpy3DPeerAsync, ::cuMemcpyPeer