Function rcudnn_sys::cudaMemcpyToSymbol

pub unsafe extern "C" fn cudaMemcpyToSymbol(
    symbol: *const c_void, 
    src: *const c_void, 
    count: usize, 
    offset: usize, 
    kind: cudaMemcpyKind
) -> cudaError_t

\brief Copies data to the given symbol on the device

Copies \p count bytes from the memory area pointed to by \p src to the memory area pointed to by \p offset bytes from the start of symbol \p symbol. The memory areas may not overlap. \p symbol is a variable that resides in global or constant memory space. \p kind can be either ::cudaMemcpyHostToDevice, ::cudaMemcpyDeviceToDevice, or ::cudaMemcpyDefault. Passing ::cudaMemcpyDefault is recommended, in which case the type of transfer is inferred from the pointer values. However, ::cudaMemcpyDefault is only allowed on systems that support unified virtual addressing.

\param symbol - Device symbol address \param src - Source memory address \param count - Size in bytes to copy \param offset - Offset from start of symbol in bytes \param kind - Type of transfer

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorInvalidSymbol, ::cudaErrorInvalidMemcpyDirection, ::cudaErrorNoKernelImageForDevice \notefnerr \note_sync \note_string_api_deprecation \note_init_rt \note_callback

\sa ::cudaMemcpy, ::cudaMemcpy2D, ::cudaMemcpy2DToArray, ::cudaMemcpy2DFromArray, ::cudaMemcpy2DArrayToArray, ::cudaMemcpyFromSymbol, ::cudaMemcpyAsync, ::cudaMemcpy2DAsync, ::cudaMemcpy2DToArrayAsync, ::cudaMemcpy2DFromArrayAsync, ::cudaMemcpyToSymbolAsync, ::cudaMemcpyFromSymbolAsync, ::cuMemcpy, ::cuMemcpyHtoD, ::cuMemcpyDtoD