Function rcudnn_sys::cudaOccupancyMaxActiveBlocksPerMultiprocessor

pub unsafe extern "C" fn cudaOccupancyMaxActiveBlocksPerMultiprocessor(
    numBlocks: *mut c_int, 
    func: *const c_void, 
    blockSize: c_int, 
    dynamicSMemSize: usize
) -> cudaError_t

\brief Returns occupancy for a device function

Returns in \p *numBlocks the maximum number of active blocks per streaming multiprocessor for the device function.

\param numBlocks - Returned occupancy \param func - Kernel function for which occupancy is calculated \param blockSize - Block size the kernel is intended to be launched with \param dynamicSMemSize - Per-block dynamic shared memory usage intended, in bytes

\return ::cudaSuccess, ::cudaErrorInvalidDevice, ::cudaErrorInvalidDeviceFunction, ::cudaErrorInvalidValue, ::cudaErrorUnknown, \notefnerr \note_init_rt \note_callback

\sa ::cudaOccupancyMaxActiveBlocksPerMultiprocessorWithFlags, \ref ::cudaOccupancyMaxPotentialBlockSize(int*, int*, T, size_t, int) "cudaOccupancyMaxPotentialBlockSize (C++ API)", \ref ::cudaOccupancyMaxPotentialBlockSizeWithFlags(int*, int*, T, size_t, int, unsigned int) "cudaOccupancyMaxPotentialBlockSizeWithFlags (C++ API)", \ref ::cudaOccupancyMaxPotentialBlockSizeVariableSMem(int*, int*, T, UnaryFunction, int) "cudaOccupancyMaxPotentialBlockSizeVariableSMem (C++ API)", \ref ::cudaOccupancyMaxPotentialBlockSizeVariableSMemWithFlags(int*, int*, T, UnaryFunction, int, unsigned int) "cudaOccupancyMaxPotentialBlockSizeVariableSMemWithFlags (C++ API)", ::cuOccupancyMaxActiveBlocksPerMultiprocessor