Function rcudnn_sys::cudaStreamGetPriority

pub unsafe extern "C" fn cudaStreamGetPriority(
    hStream: cudaStream_t, 
    priority: *mut c_int
) -> cudaError_t

\brief Query the priority of a stream

Query the priority of a stream. The priority is returned in in \p priority. Note that if the stream was created with a priority outside the meaningful numerical range returned by ::cudaDeviceGetStreamPriorityRange, this function returns the clamped priority. See ::cudaStreamCreateWithPriority for details about priority clamping.

\param hStream - Handle to the stream to be queried \param priority - Pointer to a signed integer in which the stream's priority is returned

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorInvalidResourceHandle \notefnerr \note_init_rt \note_callback

\sa ::cudaStreamCreateWithPriority, ::cudaDeviceGetStreamPriorityRange, ::cudaStreamGetFlags, ::cuStreamGetPriority