Function rcudnn_sys::cudaFuncGetAttributes

pub unsafe extern "C" fn cudaFuncGetAttributes(
    attr: *mut cudaFuncAttributes, 
    func: *const c_void
) -> cudaError_t

\brief Find out attributes for a given function

This function obtains the attributes of a function specified via \p func. \p func is a device function symbol and must be declared as a \c global function. The fetched attributes are placed in \p attr. If the specified function does not exist, then ::cudaErrorInvalidDeviceFunction is returned. For templated functions, pass the function symbol as follows: func_name<template_arg_0,...,template_arg_N>

Note that some function attributes such as \ref ::cudaFuncAttributes::maxThreadsPerBlock "maxThreadsPerBlock" may vary based on the device that is currently being used.

\param attr - Return pointer to function's attributes \param func - Device function symbol

\return ::cudaSuccess, ::cudaErrorInvalidDeviceFunction \notefnerr \note_string_api_deprecation2 \note_init_rt \note_callback

\sa \ref ::cudaFuncSetCacheConfig(const void*, enum cudaFuncCache) "cudaFuncSetCacheConfig (C API)", \ref ::cudaFuncGetAttributes(struct cudaFuncAttributes*, T*) "cudaFuncGetAttributes (C++ API)", \ref ::cudaLaunchKernel(const void *func, dim3 gridDim, dim3 blockDim, void **args, size_t sharedMem, cudaStream_t stream) "cudaLaunchKernel (C API)", ::cudaSetDoubleForDevice, ::cudaSetDoubleForHost, ::cuFuncGetAttribute